refactor dropdown and fix signbuttons
