+++ frontend/components/SignButtons/index.jsx
@@ -61,7 +61,7 @@ export const SignButtons = () => {
 
   return (
     <Container>
-      <DropDown items={dropItems} dropDownId='user-menu' toggleId='toggle-user-menu' style={{ background: theme.colors.mainBg}}>
+      <DropDown items={dropItems} dropDownId='user-menu' toggleId='toggle-user-menu'>
         <SafeImage
           src={user?.photoURL ?? defaultIcon}
           width={37}
