+++ frontend/components/SignButtons/styles.ts
@@ -1,9 +1,11 @@
 import styled, { DefaultTheme, css } from 'styled-components';
     
 export const Container = styled.div`
-  ${({ }: { theme: DefaultTheme }) => css`
+  ${({ theme }: { theme: DefaultTheme }) => css`
     display: flex;
     justify-content: space-between;
-    
+    background: ${theme.colors.mainBg};
+    padding: ${theme.spacings.xsmall};
+    border-radius: ${theme.radius.small};
   `}
 `;
